requirements.txt
discord
python-dotenv
jinja2
markdown
pymdown-extensions
nh3
emoji
gitpython
